+++ iv-outlier-detection/test/adapter.test.ts
@@ -1,5 +1,6 @@
 import { assert } from 'chai'
-import { Requester, assertSuccess, assertError } from '@chainlink/external-adapter'
+import { assertSuccess, assertError } from '@chainlink/adapter-test-helpers'
+import { Requester } from '@chainlink/external-adapter'
 import { AdapterRequest } from '@chainlink/types'
 import { execute } from '../src/adapter'
 
@@ -7,7 +8,7 @@ describe('execute', () => {
   const jobID = '1'
 
   context('successful calls @integration', () => {
-    const meta = { latestAnswer: 60 }
+    const meta = { latestAnswer: 0 }
     const multiply = 1
     const referenceContract = '0x00'
 
@@ -85,7 +86,7 @@ describe('execute', () => {
   })
 
   context('validation error', () => {
-    const meta = { latestAnswer: 60 }
+    const meta = { latestAnswer: 0 }
 
     const requests = [
       { name: 'empty body', testData: {} },
@@ -125,7 +126,7 @@ describe('execute', () => {
   })
 
   context('error calls @integration', () => {
-    const meta = { latestAnswer: 60 }
+    const meta = { latestAnswer: 0 }
     const referenceContract = '0x00'
     const multiply = 1