+++ test-helpers/src/index.ts
@@ -1,0 +1,4 @@
+import * as helpers from './helpers'
+import * as behaviors from './behaviors'
+
+export = { ...helpers, ...behaviors }